Attempted (and failed) serialization
Grabbag of changes, but go hung up on trying to implement serialization of Recipes. My thoughts are that this will make writing to / reading from a file that much easier.  This might not end up applying to the final project, so gonna take a break and wee what everyone else can make of it.